requirements.txt
qiskit>=0.39.0
opencv-python>=4.7.0
numpy>=1.22.0
Pillow>=9.3.0
matplotlib>=3.5.0
